#2: Moves db into separate package.
